% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{bayer_palettes}
\alias{bayer_palettes}
\title{Returns a ggplot-compatible palette object}
\usage{
bayer_palettes(
  name,
  n,
  all_palettes = bayer_colors,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{Name of palette that you want to use}

\item{n}{Number of values you want from the palette}

\item{all_palettes}{List of colors - generally autofilled (don't use)}

\item{type}{string, type of scale either "discrete" or "continuous"}
}
\description{
Provides palettes that correspond to official Bayer color schema
}
