% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{scale_colour_bayer_c}
\alias{scale_colour_bayer_c}
\title{Returns a ggplot-compatible palette object for continuous colors}
\usage{
scale_colour_bayer_c(name, direction = 1)
}
\arguments{
\item{name}{Name of palette that you want to use}

\item{direction}{integer: if negative, use reverse order of colors.}
}
\value{
scale
}
\description{
Provides a continuous color ggplot theme object that corresponds to official Bayer color schema
}
